refactor: extract service selection logic into a reusable ServiceForm component
