+++ _ui/src/lib/derivedcomponents/index.ts
@@ -1,5 +1,6 @@
 export { default as GenericTable } from './GenericTable.svelte';
 export { default as IncidentsTable } from './IncidentsTable.svelte';
+export { default as ServiceForm } from './ServiceForm.svelte';
 
 export type {
 	GenericTableColumn,
